transcriber_requirement.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
blinker==1.7.0
boto3==1.35.5
botocore==1.35.5
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
grpcio==1.66.0
grpcio-health-checking==1.66.0
grpcio-tools==1.66.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.8
Jinja2==3.1.4
jmespath==1.0.1
kaitaistruct==0.10
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
outcome==1.3.0.post0
protobuf==5.27.3
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
selenium==4.13.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.2
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.33.0
weaviate-client==4.7.1
websockets==13.0
wsproto==1.2.0
yt-dlp==2024.7.16
zstandard==0.23.0
